xv6 public fin

+++ xv6-public/trap.c
@@ -14,6 +14,9 @@ extern uint vectors[];  // in vectors.S: array of 256 entry pointers
 struct spinlock tickslock;
 uint ticks;
 
+int mappages(pde_t *pgdir, void *va, uint size, uint pa, int perm);
+
+
 void
 tvinit(void)
 {
@@ -54,6 +57,22 @@ trap(struct trapframe *tf)
       wakeup(&ticks);
       release(&tickslock);
     }
+
+    //current priv level : cs, user code segement selector.
+    //esp, ss in CPU internal reg
+    if(myproc()!= 0 && (tf->cs & 3) == 3){ // myproc and alarmhandler is not null.      
+      myproc()->passedticks ++;
+      if(myproc()->passedticks == myproc()->alarmticks){
+        myproc()->passedticks = 0;
+        //push eip
+        tf->esp -= 4; 
+        *((uint*)(tf->esp)) = tf->eip;
+        //change eip to alarmhandler
+        tf->eip = (uint)myproc()->alarmhandler;
+      }
+    }
+
+
     lapiceoi();
     break;
   case T_IRQ0 + IRQ_IDE:
@@ -86,6 +105,22 @@ trap(struct trapframe *tf)
               tf->trapno, cpuid(), tf->eip, rcr2());
       panic("trap");
     }
+
+    if(tf->trapno == T_PGFLT){
+      char *mem;
+      uint a;
+      a = PGROUNDDOWN(rcr2()); //rcr2 : VA which caused PF
+
+      uint newsz;
+      newsz = myproc()->sz;
+      for(; a < newsz; a += PGSIZE){
+        mem = kalloc();
+        memset(mem, 0, PGSIZE);
+        mappages(myproc()->pgdir, (char*)a, PGSIZE, V2P(mem), PTE_W|PTE_U);
+      }
+      break;
+    }
+
     // In user space, assume process misbehaved.
     cprintf("pid %d %s: trap %d err %d on cpu %d "
             "eip 0x%x addr 0x%x--kill proc\n",
@@ -110,3 +145,4 @@ trap(struct trapframe *tf)
   if(myproc() && myproc()->killed && (tf->cs&3) == DPL_USER)
     exit();
 }
+
